Move validator class in a separate folder
